modify chat_table
